# web/openapi.json
{
  "components": {
    "schemas": {
...
      "BookmarkCreateIn": {
        "properties": {
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "생략 시 내용 해시로 생성",
            "title": "Id"
          },
          "link": {
            "default": "",
            "title": "Link",
            "type": "string"
          },
          "meta": {
            "additionalProperties": true,
            "description": "구조화 메타 — 제안서는 {task_id, task_key, article_ids, matching_version, prompt_version}",
            "title": "Meta",
            "type": "object"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "news | proposal | opportunity | task",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title"
        ],
        "title": "BookmarkCreateIn",
        "type": "object"
      },
      "BookmarkOut": {
        "description": "북마크 응답 (store.bookmarks.Bookmark).",
        "properties": {
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "decided_at": {
            "default": "",
            "title": "Decided At",
            "type": "string"
          },
          "decision_note": {
            "default": "",
            "title": "Decision Note",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "link": {
            "default": "",
            "title": "Link",
            "type": "string"
          },
          "meta": {
            "additionalProperties": true,
            "description": "구조화 메타(제안서 근거 관계 등)",
            "title": "Meta",
            "type": "object"
          },
          "status": {
            "default": "pending",
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "user_id": {
            "default": "local",
            "title": "User Id",
            "type": "string"
          },
          "workspace_id": {
            "default": "default",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "title"
        ],
        "title": "BookmarkOut",
        "type": "object"
      },
      "BookmarkStatusIn": {
        "properties": {
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "status": {
            "description": "pending | adopted | rejected",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "BookmarkStatusIn",
        "type": "object"
      },
      "BookmarkUpdateIn": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "BookmarkUpdateIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "조선소 작업정의 × 외부 기술동향 매칭 — React 전환용 백엔드 계약",
    "title": "InsightBoard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/board/brief": {
      "get": {
        "operationId": "board_brief_api_board_brief_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 30,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "캐시 무시 재생성",
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "description": "캐시 무시 재생성",
              "title": "Force",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Board Brief Api Board Brief Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Board Brief",
        "tags": [
          "board"
        ]
      }
    },
    "/api/board/digest": {
      "get": {
        "description": "오늘의 개인화 다이제스트 — 기사 3~5건 + '왜 내 업무와 관련 있는가'.\n\n랭킹·이유는 links(저장된 매칭)와 페르소나의 규칙 조합(LLM 미사용,\nstore/rank.py). '관련 없음'(dismiss) 처리한 기사는 제외된다.",
        "operationId": "board_digest_api_board_digest_get",
        "parameters": [
          {
            "description": "다이제스트 대상 기간",
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 3,
              "description": "다이제스트 대상 기간",
              "maximum": 30,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 12,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Board Digest Api Board Digest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Board Digest",
        "tags": [
          "board"
        ]
      }
    },
    "/api/bookmarks": {
      "get": {
        "operationId": "list_bookmarks_api_bookmarks_get",
        "parameters": [
          {
            "description": "타입 필터(news/proposal/...)",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "타입 필터(news/proposal/...)",
              "title": "Type"
            }
          },
          {
            "description": "상태 필터(pending/adopted/rejected)",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "상태 필터(pending/adopted/rejected)",
              "title": "Status"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookmarkOut"
                  },
                  "title": "Response List Bookmarks Api Bookmarks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Bookmarks",
        "tags": [
          "bookmarks"
        ]
      },
      "post": {
        "operationId": "create_bookmark_api_bookmarks_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Bookmark",
        "tags": [
          "bookmarks"
        ]
      }
    },
...
    "/api/bookmarks/{bm_id}": {
      "delete": {
        "operationId": "delete_bookmark_api_bookmarks__bm_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "bm_id",
            "required": true,
            "schema": {
              "title": "Bm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Bookmark Api Bookmarks  Bm Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Bookmark",
        "tags": [
          "bookmarks"
        ]
      },
      "patch": {
        "operationId": "update_bookmark_api_bookmarks__bm_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "bm_id",
            "required": true,
            "schema": {
              "title": "Bm Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Bookmark",
        "tags": [
          "bookmarks"
        ]
      }
    },
    "/api/bookmarks/{bm_id}/status": {
      "post": {
        "operationId": "set_status_api_bookmarks__bm_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "bm_id",
            "required": true,
            "schema": {
              "title": "Bm Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkStatusIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Status",
        "tags": [
          "bookmarks"
        ]
      }
    },
...
  }
}